Refactor, merge SSP case study
