+++ main.py
@@ -1,5 +1,4 @@
 import numpy as np
-
 import data.datasets as datasets
 from helpers.plotters import *
 from helpers.funcs import *
